requirements.txt
# Core LangChain packages
langchain==1.0.2
langchain-core==1.0.1
langchain-openai==1.0.1
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-classic==1.0.0

# OpenAI
openai>=1.12.0
tiktoken>=0.7.0

# PDF Processing
pymupdf>=1.23.0

# Vector Store
faiss-cpu>=1.7.4

# Environment Variables
python-dotenv>=1.0.0

# Additional dependencies (automatically installed with above packages)
pydantic>=2.0.0
numpy>=1.26.0
